+++ xbmc/utils/JobManager.h
@@ -326,7 +326,7 @@ protected:
    \return true if the job has been cancelled, else returns false.
    \sa IJobCallback, CJob
    */
-  bool  OnJobProgress(unsigned int progress, unsigned int total, const CJob *job) const;
+  bool  OnJobProgress(uint64_t progress, uint64_t total, const CJob *job) const;
 
 private:
   // private construction, and no assignements; use the provided singleton methods
